fix : pk 다루는 부분 int -> long 으로 수정

+++ src/main/java/com/avg/lawsuitmanagement/member/controller/MemberController.java
@@ -64,12 +64,12 @@ public class MemberController {
     }
 
     @GetMapping("/employees/{employeeId}")
-    public ResponseEntity<MemberDtoNonPass> getEmployee(@PathVariable int employeeId) {
+    public ResponseEntity<MemberDtoNonPass> getEmployee(@PathVariable long employeeId) {
         return ResponseEntity.ok(memberService.getMemberInfoById(employeeId));
     }
 
     @PutMapping("/employees/{employeeId}")
-    public ResponseEntity<Void> updateEmployee(@PathVariable int employeeId, @RequestBody @Valid
+    public ResponseEntity<Void> updateEmployee(@PathVariable long employeeId, @RequestBody @Valid
         MemberUpdateForm form) {
         memberService.updateMemberInfo(employeeId, form);
         return ResponseEntity.ok().build();
